fix: 좋아요 기능 like Controller 처리
